+++ src/App.css
@@ -39,9 +39,18 @@ color: rgb(18, 36, 50);
 }
 
 .mainPic {
+  padding-left: 10px;
   max-width: 400px;
   width: 100%;
   border-radius: 50px;
+}
+
+@media (min-width: 800px) {
+  .mainPic {
+    max-width: 400px;
+    width: 100%;
+    border-radius: 50px;
+  }
 }
 
 .header {
@@ -94,18 +103,36 @@ a {
   text-decoration: none;
 }
 
-h1, h2 {
+h1 {
   font-weight: 700;
-  font-size: 50px;
-  line-height: 60px;
-  margin-bottom: 14px;
+  font-size: 40px;
   pointer-events: none;
+}
+
+@media (min-width: 800px) {
+  h1 {
+      font-size: 40px;
+  }
+}
+
+.pad1 {
+  padding: 0 10px;
 }
 
 .title {
   color: transparent;  
   text-shadow: 0 0 0 white;
   background: rgb(18, 36, 50);
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 10px;
+  padding: 10px;
+}
+
+.title2 {
+  color: transparent;  
+  text-shadow: 0 0 0 white;
 }
 
 #canvas {
@@ -113,14 +140,6 @@ h1, h2 {
   top: 0;
   left: 0;
   z-index: 0;
-}
-
-@media (min-width: 800px) {
-  h1, h2 {
-      font-size: 70px;
-      line-height: 78px;
-      margin-bottom: 20px;
-  }
 }
 
 h3 {
@@ -152,6 +171,7 @@ p {
   display: flex;
   flex-direction: row;
   justify-content: left;
+  padding-left: 10px;
 }
 
 @media (min-width: 800px) {
@@ -161,12 +181,9 @@ p {
   }
 }
 
-
-
 .container1 {
   max-width: 1080px;
   width: 100%;
-
 }
 
 .row {
@@ -180,7 +197,7 @@ p {
 @media (min-width: 800px) {
   .row {
       flex-direction: row;
-      padding-right: 20px;
+
       gap: 20px;
       align-items: center;
   }
@@ -213,6 +230,12 @@ p {
   line-height: 28px;
   font-weight: 600;
   margin-bottom: 4px;
+}
+
+.footerLabel {
+  color: rgba(18, 36, 50, 0.48);
+  font-size: 17px;
+  font-weight: 600;
 }
 
 @media (min-width: 800px) {
