+++ src/saves.txt
@@ -1,0 +1,15 @@
+      <FirstTopic foto={foto} mainOpacity={mainOpacity}/>
+      <SecondTopic foto={foto} mainOpacity={mainOpacity}/>
+      <ThirdTopic foto={foto} mainOpacity={mainOpacity}/>
+      <ComercialProjects foto={foto} mainOpacity={mainOpacity}/>
+
+
+
+
+
+      <br/><br/><br/><br/><br/><br/><br/>
+      <div style={{background: "orange", height: "50px", width: "200px", padding: "10px"}}>
+        <img src={skillsPic}/>
+
+
+      </div>
